refactored listenToMany
