+++ src/index.js
@@ -148,7 +148,29 @@ exports.ListenerMixin = require('./ListenerMixin');
 
 exports.listenTo = require('./listenTo');
 
-exports.listenToMany = require('./listenToMany');
+exports.listenToMany = function(listenables){
+    return {
+        /**
+         * Set up the mixin before the initial rendering occurs. Import methods from `ListenerMethods`
+         * and then make the call to `listenTo` with the arguments provided to the factory function
+         */
+        componentDidMount: function() {
+            for(var m in Reflux.ListenerMethods){
+                if (this[m] !== Reflux.ListenerMethods[m]){
+                    if (this[m]){
+                        throw "Can't have other property '"+m+"' when using Reflux.listenToMany!";
+                    }
+                    this[m] = Reflux.ListenerMethods[m];
+                }
+            }
+            this.listenToMany(listenables);
+        },
+        /**
+         * Cleans up all listener previously registered.
+         */
+        componentWillUnmount: Reflux.ListenerMethods.stopListeningToAll
+    };
+};
 
 exports.joinTrailing = exports.all = maker("last"); // Reflux.all alias for backward compatibility
